+++ lib/hexapdf/document/layout.rb
@@ -639,15 +639,15 @@ module HexaPDF
       # If the +properties+ hash is not empty, the retrieved style is duplicated and the properties
       # hash is applied to it.
       #
-      # Finally, a default font is set if necessary to ensure that the style object works in all
-      # cases.
+      # Finally, a default font (the one from the :base style or otherwise 'Times') is set if
+      # necessary to ensure that the style object works in all cases.
       def retrieve_style(style, properties = nil)
         if style.kind_of?(Symbol) && !@styles.key?(style)
           raise HexaPDF::Error, "Style #{style} not defined"
         end
         style = HexaPDF::Layout::Style.create(@styles[style] || style || @styles[:base])
         style = style.dup.update(**properties) unless properties.nil? || properties.empty?
-        style.font('Times') unless style.font?
+        style.font(@styles[:base].font? && @styles[:base].font || 'Times') unless style.font?
         unless style.font.respond_to?(:pdf_object)
           name, options = *style.font
           style.font(@document.fonts.add(name, **(options || {})))