+++ CHANGELOG.md
@@ -16,6 +16,8 @@
   fall back to Times and ZapfDingbats if necessary
 * [HexaPDF::Document::Layout#table_box] to merge the `:cell` keys that define
   the cell style instead of using the last one
+* [HexaPDF::Document::Layout] style retrieval to fall back to using the font of
+  the `:base` style and only if that doesn't exist to 'Times'
 
 
 ## 0.37.2 - 2024-02-27